I think its getting there.
I preserved these comments by @waalge from one of the squashed commits:

- Its easier to separate content more into paragraphs.
Even if a paragraph is short.
For example, for the blockchain section we can do a paragraph on
bitcoin, and then a comparator on cardano.
Hopefully that is demonstrated.

- The section on blockchain should _not_ have BLN in it.
Instead it should be referenced in later sections, since the content has
now been introduced.

- Please get a spell checker

- Notes indented with `>` are notes, right?
I think style wise, we should keep it "somewhat, but overly, formal".
Its not a paper, but its also not a transcript of us trying to explain
it to someone.